# docs/swagger.yaml
definitions:
  entity.AIAnalyticsRequest:
    properties:
      deep_work:
        $ref: '#/definitions/entity.DeepWorkData'
      domains:
        items:
          type: string
        minItems: 1
        type: array
      domains_count:
        minimum: 1
        type: integer
      engagement_rate:
        maximum: 100
        minimum: 0
        type: number
      period:
        type: string
      timestamp:
        type: string
      tracked_hours:
        minimum: 0
        type: number
      user_id:
        type: string
    required:
    - deep_work
    - domains
    - domains_count
    - engagement_rate
    - tracked_hours
    type: object
...
  entity.DeepWorkData:
    properties:
      average_minutes:
        description: средняя длительность сессии
        type: number
      deep_work_rate:
        description: '% от tracked time'
        type: number
      longest_minutes:
        description: самая длинная сессия
        type: number
      sessions_count:
        description: количество deep work сессий
        type: integer
      top_domains:
        description: топ доменов для deep work
        items:
          $ref: '#/definitions/entity.DeepWorkDomain'
        type: array
      total_hours:
        description: общее время в часах
        type: number
      total_minutes:
        description: общее время в deep work
        type: number
    type: object
  entity.DeepWorkDomain:
    properties:
      domain:
        type: string
      minutes:
        type: number
      sessions:
        type: integer
    type: object
  entity.DetailedAnalysis:
    properties:
      behavior_insights:
        items:
          type: string
        type: array
      domain_breakdown:
        $ref: '#/definitions/entity.DomainBreakdown'
      key_findings:
        items:
          type: string
        type: array
      productivity_score:
        $ref: '#/definitions/entity.ProductivityScore'
    type: object
  entity.DomainAnalysis:
    properties:
      analysis:
        $ref: '#/definitions/entity.DetailedAnalysis'
      focus_insight:
        type: string
      focus_level:
        type: string
      recommendations:
        items:
          type: string
        type: array
      work_pattern:
        type: string
    type: object
  entity.DomainBreakdown:
    properties:
      communication:
        description: Gmail, Telegram, etc.
        items:
          type: string
        type: array
      development:
        description: localhost, CodeSandbox, etc.
        items:
          type: string
        type: array
      distractions:
        description: YouTube, social, etc.
        items:
          type: string
        type: array
      research:
        description: Stack Overflow, docs, etc.
        items:
          type: string
        type: array
      work_tools:
        description: GitHub, Jira, etc.
        items:
          type: string
        type: array
    type: object
...
  entity.FocusLevelResponse:
    properties:
      focus_level:
        type: string
      insight:
        type: string
      method:
        description: '"ai" или "fallback"'
        type: string
      timestamp:
        type: string
    type: object
...
  entity.ProductivityScore:
    properties:
      balance:
        description: work/life domains
        type: integer
      efficiency:
        description: engagement rate
        type: integer
      explanation:
        description: краткое объяснение счета
        type: string
      focus:
        description: на основе deep work
        type: integer
      overall:
        description: 0-100
        type: integer
    type: object
...
  wrapper.ErrorWrapper:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  wrapper.ResponseWrapper:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /ai-analytics/domain-usage:
    post:
      consumes:
      - application/json
      description: Get AI-powered analysis of user's domain usage patterns, productivity
        insights, and recommendations
      parameters:
      - description: Analytics request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.AIAnalyticsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.DomainAnalysis'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Analyze domain usage with AI
      tags:
      - /api/v1/admin/ai-analytics
  /ai-analytics/focus-level:
    get:
      consumes:
      - application/json
      description: Get basic focus level assessment based on domain count (fallback
        method)
      parameters:
      - description: Number of unique domains
        in: query
        name: domains_count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wrapper.ResponseWrapper'
            - properties:
                data:
                  $ref: '#/definitions/entity.FocusLevelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wrapper.ErrorWrapper'
      summary: Get focus level without AI
      tags:
      - /api/v1/admin/ai-analytics
...
swagger: "2.0"